Backend/requirements.txt
llama_index.core
llama_index.llms.ollama
llama_index.embeddings.huggingface
llama-index-readers-file
flask
requests
flask_cors